More cleanup and refactor
