+++ crates/cubecl-runtime/src/throughput/mod.rs
@@ -1,9 +1,6 @@
-mod base;
 mod benchmarker;
 mod cache;
 mod cmma;
-
-pub use base::*;
 pub use benchmarker::*;
 pub use cache::*;
 pub use cmma::*;
