+++ crates/cubecl-std/src/throughput/runners/memory_direct.rs
@@ -1,48 +1,46 @@
 use cubecl::prelude::*;
 use cubecl_core as cubecl;
-use cubecl_runtime::throughput::{KernelConfig, LaunchConfig, ThroughputKey, ThroughputRunner};
+use cubecl_runtime::throughput::{KernelConfig, ThroughputKey};
 
-pub struct MemoryDirectRunner;
+use crate::throughput::LaunchConfig;
 
-impl<R: Runtime> ThroughputRunner<R> for MemoryDirectRunner {
-    fn build_kernel(
-        client: &ComputeClient<R>,
-        key: ThroughputKey,
-        config: LaunchConfig,
-    ) -> KernelConfig {
-        let client = client.clone();
-        let dtype = key.dtype;
+pub fn build_kernel<R: Runtime>(
+    client: &ComputeClient<R>,
+    key: ThroughputKey,
+    config: LaunchConfig,
+) -> KernelConfig {
+    let client = client.clone();
+    let dtype = key.dtype;
 
-        let line_bytes = config.vector_size * dtype.size();
+    let line_bytes = config.vector_size * dtype.size();
 
-        const TARGET_BYTES: usize = 256 * 1024 * 1024;
-        let max_alloc = client.properties().memory.max_page_size as usize;
-        let target = TARGET_BYTES.min(max_alloc);
+    const TARGET_BYTES: usize = 256 * 1024 * 1024;
+    let max_alloc = client.properties().memory.max_page_size as usize;
+    let target = TARGET_BYTES.min(max_alloc);
 
-        let total_threads = config.cube_count * config.cube_dim;
-        let num_lines = (target / line_bytes).max(total_threads);
-        let bytes = num_lines * line_bytes;
+    let total_threads = config.cube_count * config.cube_dim;
+    let num_lines = (target / line_bytes).max(total_threads);
+    let bytes = num_lines * line_bytes;
 
-        let kernel = Box::new(move |iterations: usize| unsafe {
-            let in_handle = client.empty(bytes);
-            let out_handle = client.empty(bytes);
+    let kernel = Box::new(move |iterations: usize| unsafe {
+        let in_handle = client.empty(bytes);
+        let out_handle = client.empty(bytes);
 
-            memory_direct_throughput::launch_unchecked(
-                &client,
-                CubeCount::Static(config.cube_count as u32, 1, 1),
-                CubeDim::new_1d(config.cube_dim as u32),
-                config.vector_size,
-                BufferArg::from_raw_parts(in_handle, num_lines),
-                BufferArg::from_raw_parts(out_handle, num_lines),
-                iterations,
-                dtype.into(),
-            )
-        });
+        memory_direct_throughput::launch_unchecked(
+            &client,
+            CubeCount::Static(config.cube_count as u32, 1, 1),
+            CubeDim::new_1d(config.cube_dim as u32),
+            config.vector_size,
+            BufferArg::from_raw_parts(in_handle, num_lines),
+            BufferArg::from_raw_parts(out_handle, num_lines),
+            iterations,
+            dtype.into(),
+        )
+    });
 
-        let ops_count = 2 * num_lines * config.vector_size;
+    let ops_count = 2 * num_lines * config.vector_size;
 
-        KernelConfig { kernel, ops_count }
-    }
+    KernelConfig { kernel, ops_count }
 }
 
 #[cube(launch_unchecked)]
