+++ crates/cubecl-std/src/throughput/runners/mod.rs
@@ -1,7 +1,3 @@
-mod compute_cmma;
-mod compute_direct;
-mod memory_direct;
-
-pub(crate) use compute_cmma::*;
-pub(crate) use compute_direct::*;
-pub(crate) use memory_direct::*;
+pub mod compute_cmma;
+pub mod compute_direct;
+pub mod memory_direct;
